Dissect nucypher keyring and utils.
